- warning visual studio 2013 64 (3/?)

+++ include/general/visual_mainwindow.h
@@ -30,7 +30,7 @@ public:
     cMMSpecArg  Arg()       { return mArg;    }
     int         Type();
     vector < pair < int, QWidget*> >    Widgets()    { return vWidgets; }
-    int         NbWidgets() { return vWidgets.size(); }
+    int         NbWidgets() { return (int)vWidgets.size(); }
 
 private:
     cMMSpecArg  mArg;
